Change the Principal entitie and the PermissionsInfo command
The PermissionsInfo command returns a Principal[] instead of
Permission[], so I change the Command. But the Principal returned by
Adobe Connect doesn't have the firstName and lastName when the Principal
is an user. Because of that the Principal entite was modified to set the
firstName and lastName when only the name is setted.